code cleanup and comments
